Move std.map functions into each typeclass companion object with JvmStatic annotation and remove imports from deprecations
