+++ arrow-libs/optics/arrow-optics/src/main/kotlin/arrow/optics/Traversal.kt
@@ -257,6 +257,13 @@ interface PTraversal<S, T, A, B> : PTraversalOf<S, T, A, B> {
             FA.run { s.traverse(FA, f).map { it.fix() } }
           }
       }
+
+    @JvmStatic
+    fun <K, V> map(): Traversal<Map<K, V>, V> = object : Traversal<Map<K, V>, V> {
+      override fun <F> modifyF(FA: Applicative<F>, s: Map<K, V>, f: (V) -> Kind<F, V>): Kind<F, Map<K, V>> = FA.run {
+        s.k().traverse(FA, f)
+      }
+    }
   }
 
   /**
