+++ arrow-libs/optics/arrow-optics/src/main/kotlin/arrow/optics/extensions/map/filterIndex/MapKFilterIndex.kt
@@ -15,7 +15,7 @@ import arrow.optics.extensions.MapKFilterIndex
   "arrow.optics.extensions package is being deprecated. Use the exposed function in the instance for Map from the companion object of the typeclass instead.",
   ReplaceWith(
     "FilterIndex.map<K, V>().filter(p)",
-    "arrow.optics.map", "arrow.optics.typeclasses.FilterIndex"
+    "arrow.optics.typeclasses.FilterIndex"
   ),
   DeprecationLevel.WARNING
 )
