+++ arrow-libs/optics/arrow-optics/src/main/kotlin/arrow/optics/extensions/mapk/index/MapKIndex.kt
@@ -23,7 +23,7 @@ internal val index_singleton: MapKIndex<Any?, Any?> = object : MapKIndex<Any?, A
   "arrow.optics.extensions package is being deprecated, and it will be removed in 0.13.",
   ReplaceWith(
     "this compose Index.map<K, V>().index(i)",
-    "arrow.optics.map", "arrow.optics.typeclasses.Index", "arrow.optics.compose"
+    "arrow.optics.typeclasses.Index", "arrow.optics.compose"
   ),
   level = DeprecationLevel.WARNING
 )
@@ -43,7 +43,7 @@ fun <K, V, T> PLens<T, T, MapK<K, V>, MapK<K, V>>.index(i: K): POptional<T, T, V
   "arrow.optics.extensions package is being deprecated, and it will be removed in 0.13.",
   ReplaceWith(
     "this compose Index.map<K, V>().get(i)",
-    "arrow.optics.map", "arrow.optics.typeclasses.Index", "arrow.optics.compose"
+    "arrow.optics.typeclasses.Index", "arrow.optics.compose"
   ),
   level = DeprecationLevel.WARNING
 )
@@ -60,7 +60,7 @@ operator fun <K, V, T> PLens<T, T, MapK<K, V>, MapK<K, V>>.get(i: K): POptional<
   "MapK is being deprecated. Use the instance for Map from the companion object of the typeclass.",
   ReplaceWith(
     "Index.map<K, V>()",
-    "arrow.optics.map", "arrow.optics.typeclasses.Index"
+    "arrow.optics.typeclasses.Index"
   ),
   DeprecationLevel.WARNING
 )
