+++ arrow-libs/optics/arrow-optics/src/test/kotlin/arrow/optics/std/MapTest.kt
@@ -8,7 +8,6 @@ import arrow.core.test.generators.functionAToB
 import arrow.core.test.generators.genSetK
 import arrow.core.test.generators.mapK
 import arrow.optics.Iso
-import arrow.optics.mapToSet
 import arrow.optics.test.laws.IsoLaws
 import arrow.optics.toSetK
 import arrow.typeclasses.Eq
